Revert "feat: get statictis for guide and tour"
This reverts commit d77bee767d7629cfb512b703326de383aba784fa.

+++ src/main/java/com/example/localguidebe/controller/GuideController.java
@@ -44,7 +44,7 @@ public class GuideController {
                 new SearchGuideDTO(
                     guides.stream().map(userToGuideDtoConverter::convert).toList(),
                     guides.getTotalPages(),
-                    (int) guides.getTotalElements())));
+                    guides.getTotalElements())));
   }
 
   @GetMapping("/{id}")
